Baseline Modularization
The latent_semantic_indexing function returns a numpy array for
Procrustes Analysis, and works as intended as of now.

LDA and/or doc2vec and then BERT and/or GPT is pending; the code just needs its respective
functions to be implemented, as the code is quite modular and reusable
now.

Procrustes Analysis is working, though the appending zeros functionality
is pending. I want to implement another distributional semantic model;
I think that takes priority.

Docstrings have been written for each function. They will be refined
later, but have the most important information needed.