Universal document-feature matrix
Will take in a vectorized corpus, and convert it to a document-feature
matrix. IMPORTANT: it will append and fill zeros to match the
user-defined number of documents and topics as the rows and columns.

This is important because the Procrustes analysis requires both matrices
to be the same size. It also mentions that if they aren't the same size,
just append zeros.