Created universal vectorize function and much more
Created a universal vectorize function that works on both LSI and LDA.

This also removes the funky boolean parameters for both. That did work, but I didn't like it because it made the code less modular.

Has new code in dunder main to show how to use these revamped functions. You have to use more functions to do the same task, but you have more granular control of what you want to do, and it will be easier to add new models to train in this file.

Added docstrings to the new functions.