Updated document selection function.
The document selection function selected documents just starting from the beginning, which was no good because it was training the models using the test documents, as well as treated the whole reuters corpus as the same.

Instead, I modified this function to select only the training documents starting from the first training document, position 3020 in the Reuters corpus (3019 counting from 0).