Output models to respective folders
Along with folder creation when doesn't exist, the document feature matrices will be outputted to its respective folders.

To add more models to this, simply add strings to the models list at the top of the function.

Do note that the model_type string argument and model_folders entries need to match case.