Implemented Save Models Functionctionality
This takes a similar form of saving the document-feature matrices by creating multiple folders for each model type. But instead of saving the models in that folder, they are saved in another folder (timestamped), since the models are saved as two files.

This is the folder hierarchy, with fill-in words:

saved_models/model/timestamp/model_files.