Update procrustes_analysis.py
Added first version of load model function. More in-depth testing needs to be done, but the loaded model will produce the same standardized matrix generated by Procrustes analysis. This has been tested using the online text compare utility.

load_model takes in two parameters, the model_type and model_timestamp. model_type is a string that determines the model to load, and model_timestamp determines the specific model to load in that folder.