Changed load_model method
Instead of entering the very specific and error-prone timestamp to load a model, a simple integer can be used.

This integer directly corresponds to the location in the model folder, whether it's lsi, lda, etc. Because of the ISO 8601 format, the saved models will always stay in that particular order.