Intermediary commit for new ProcrustesBundle class
I created class to bundle Procrustes-related variables together, aptly called "ProcrustesBundle."

The primary purpose for this is less confusion as to what a variable is for. You must explicitly use a ProcrustesBundle instantiation's getters and setters to access that data.

This grouping of related variables is both appropriate and necessary for this project, given how many variables are being tossed around. For example, instead of having to remember the many arguments that modified_procrustes_analysis takes, all you have to feed it is a ProcrustesBundle object.

I'll also be looking into giving document-feature matrices this treatment, as those objects have more data than just the matrix themselves.

With all this positive talk, it is possible that major problems could arise. I don't know the future. In any case, I'll be ready for a reversion to the looser structure if need be.