Added microseconds to ISO 8601 timestamps
This allows for the filenames to be much more unique. When running the analysis in a for loop, the analysis wouldn't be able to keep generating unique times for Python to save to.

This extra bit of precision practically removes this error. Chances are that analysis wouldn't happen faster than a single microsecond.